Rename edit action dialog to be less confusing
